@@ -76,11 +76,11 @@
     "@kkt/ncc": "~1.0.8",
     "@kkt/raw-modules": "~7.1.1",
     "@kkt/scope-plugin-options": "~7.1.1",
-    "@types/react": "~17.0.39",
-    "@types/react-dom": "~17.0.11",
-    "@types/react-test-renderer": "~17.0.1",
-    "@uiw/react-github-corners": "~1.5.3",
-    "@uiw/react-markdown-preview": "^4.0.4",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
+    "@types/react-test-renderer": "^18.0.0",
+    "@uiw/react-github-corners": "^1.5.14",
+    "@uiw/react-markdown-preview": "^4.0.6",
     "@uiw/react-loader": "^4.14.2",
     "@wcj/dark-mode": "^1.0.13",
     "code-example": "^3.3.1",
@@ -89,9 +89,9 @@
     "kkt": "~7.1.5",
     "lint-staged": "~12.3.4",
     "prettier": "~2.6.0",
-    "react": "~17.0.2",
-    "react-dom": "~17.0.2",
-    "react-test-renderer": "~17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-test-renderer": "^18.1.0",
     "tsbb": "~3.7.0"
   },
   "eslintConfig": {
